Resize and center minigame
Also reoarder logic
