+++ project.godot
@@ -18,7 +18,6 @@ config/icon="res://icon.png"
 [display]
 
 window/size/resizable=false
-window/size/fullscreen=true
 window/stretch/mode="2d"
 window/stretch/aspect="keep"
