+++ scenes/Room.tscn
@@ -13,5 +13,5 @@ scale = Vector2( 0.850781, 0.659896 )
 texture = ExtResource( 1 )
 
 [node name="Sunflowers" parent="." instance=ExtResource( 2 )]
-position = Vector2( 75, 300 )
+position = Vector2( 81, 285 )
 scale = Vector2( 0.5, 0.5 )
