+++ scenes/Room_Left.tscn
@@ -3,3 +3,6 @@
 [ext_resource path="res://scenes/Room.tscn" type="PackedScene" id=1]
 
 [node name="Room_Left" instance=ExtResource( 1 )]
+
+[node name="Sunflowers" parent="." index="1"]
+scale = Vector2( 1, 1 )
